packages/server/demo-worker/requirements.txt
livekit-agents>=0.11.0
livekit-plugins-openai>=0.10.5
livekit-plugins-deepgram>=0.6.9
livekit-plugins-silero>=0.7.3
python-dotenv~=1.0